[Fix] Use the correct window with `ejmr-counsel-file-register`
This patch wraps the body of the command's action with the
`with-ivy-window` macro so that the command always opens a file
register in the window where I used the command to begin with.

Additionally the patch improves the function by sorting the names of
the file registers and storing the history of previous selections
separate from the general Ivy and Counsel command history.