[Feature] Load the RPN Calculator package
This binds it to my `C-c x` hydra for various commands.

On a completely unrelated note, the patch also adds a custom function
to (barely) automate the process of updating packages.  Yeah---it
should be part of a separate commit, but the command doesn't work
fully the way I want anyway, so... See ya later, commit practices!