[Refactor] Rename my maps for help and window commands
The names `super-1-map` and `super-w-map` explicitly imply which keys
are bound to those maps.  But in the future I may use different prefix
keys, at which point the names will make no sense.  So this patch
renames them to `ejmr-help-map` and `ejmr-window-map`, respectively,
so that if I ever change the prefixes the map names will still be
sensible.