[Feature] Install `git-complete`
This package is currently not in any package repository so I am using
a local copy.  The patch also renames the page for completion via
Company to simply a page for any packages related to completion.  And
it binds `git-complete` to the `M-s-/` key so the input is similar to
others related to completion.