[Feature] Install and enable `auto-minor-mode`
This patch enables the package so that I can automatically enable
minor modes based on filenames and/or their contents.  I intentionally
and explicitly use `:defer nil` so that the `:minor` and
`:magic-minor` keywords are available to `use-package` elsewhere
throughout my configuration, which also means that...

    (use-package auto-minor-mode :defer nil)

...needs to remain near the top of my configuration.

This patch simply installs the package though.  I have not yet decided
which minor modes I want to enable via this mechanism, only that I
know I will be doing so in the future and as I re-evaluate the minor
modes I use in general.