[Feature] Bind `C-c C-y` to toggle Lispy for Lisp modes
I can already use the `C-c n` hydra to turn Lispy on or off for a
buffer but that brings up the entire minor mode hydra.  This patch
binds a specific key for Lispy mode in Emacs Lisp Mode and regular
Lisp Mode, because I find myself wanting to toggle Lispy most often in
those modes, and so having a dedicated key that is *barely* easier to
use than the minor mode hydra is a really, very, extremely small
convenience.