[Style] Only use Indent Guide Mode in programming modes
This patch globally disables Indent Guide Mode and replaces that
global setting with a hook for `prog-mode` (and thus any derived
modes) which enables the Indent Guide.  I find that when I am not
reading or writing code that I do not really care about seeing levels
of indentation, e.g. it does not help me when writing articles.