[Feature] Dired: Bind `* R` to a command to open READMEs
This patch creates a custom command for use in Dired, bound to `* R`
so that it will not conflict with existing commands, which will open
the README file in the current directory in View Mode and make it the
sole window.  This helps me quickly open READMEs for many projects.