[Feature] use-package: Install a handler for the `:chords` keyword
This patch installs an extension for use-package which allows me to
use a `:chords` keyword in the same way I use `:bind` throughout my
configuration.  However, for now I have disabled the extension because
it (apparently) uses the `key-chord` function.  I prefer to use
`key-seq`, which itself is a utility built atop the `key-chord`
package.  I do not want to migrate my chord bindings to use the
`:chords` keyword until I modifier the handler to use `key-seq`.