[Feature] Add a hydra for cycling word case, hyphenation, and quotes
This patch installs the `string-inflection` package which I can use
on the word under point to cycle through the forms:

1. foo_bar
2. FOO_BAR
3. FooBar
4. fooBar
5. foo-bar
6. foo_bar

This is most useful when I am working with code.  Emacs' built-in
functions for case, e.g. `capitalize-word`, and the `cycle-quotes`
package are similar to `string-inflection` in purpose.  Therefore this
patch rebinds `M-c` from the default capitalization binding to my new
hydra for using all three of these sets of commands.